ogma-core: Move diagram parsing definitions into separate modules. Refs #403.
The module `ogma-core:Command.CommonDiagram` has too many
responsibilities. Specifically, it contains a type for `Diagram`,
functions to parse diagrams from files in different formats, and
functions to formally analyze diagrams. This contrasts with how specs
are being treated inside Ogma, where there is a module that defines
specs, a separate module for spec analysis, and a separate one for spec
parsing. Having a more uniform interface for all the objects that Ogma
treats would also facilitate expanding functions that currently are only
available for diagrams or specs to apply to both. Offering cohesive
modules that embrace the single responsibility principle would
facilitate updates, development and debugging, and will help better
accommodate the needs in upcoming versions of Ogma.

This commit moves the functionality for parsing diagrams from
`Command.CommonDiagram` into several smaller modules. Specifically, a
new module is introduced for each file format supported, and an
additional new module `Data.Diagram.Parser` acts as a dispatcher or
top-level diagram loading entry point for all formats available.
