+++ ogma-core/src/Command/CommonDiagram.hs
@@ -1,4 +1,3 @@
-{-# LANGUAGE OverloadedStrings #-}
 -- Copyright 2024 United States Government as represented by the Administrator
 -- of the National Aeronautics and Space Administration. All Rights Reserved.
 --
@@ -18,254 +17,20 @@
 --
 -- | Common functionality related to diagrams.
 module Command.CommonDiagram
-    ( DiagramFormat(..)
-    , readDiagram
-    , analyzeDiagram
+    ( analyzeDiagram
     , AnalysisResult(..)
     )
   where
 
 -- External imports
-import           Control.Monad                     (when, void)
-import qualified Copilot.Core                      as Core
-import           Data.ByteString.Lazy              (toStrict)
-import qualified Data.ByteString.Lazy              as B
-import           Data.Either                       (isLeft)
-import           Data.Functor.Identity             (Identity)
-import           Data.GraphViz                     (graphEdges)
-import qualified Data.GraphViz                     as G
-import qualified Data.GraphViz.Attributes.Complete as Attributes
-import           Data.GraphViz.Commands.IO         (toUTF8)
-import qualified Data.GraphViz.Parsing             as G
-import           Data.GraphViz.PreProcessing       (preProcess)
-import qualified Data.GraphViz.Types.Generalised   as Gs
-import           Data.List                         (intercalate)
-import qualified Data.Set                          as Set
-import           Data.Text                         (Text)
-import qualified Data.Text                         as T
-import qualified Data.Text.Encoding                as T
-import qualified Data.Text.Lazy                    as LT
-import           Data.Void                         (Void)
-import           Text.Megaparsec                   (ErrorFancy (ErrorFail),
-                                                    ParsecT, choice, empty,
-                                                    errorBundlePretty,
-                                                    fancyFailure, many,
-                                                    manyTill, noneOf, parse,
-                                                    (<|>))
-import           Text.Megaparsec.Char              (alphaNumChar, char,
-                                                    digitChar, newline, space1,
-                                                    string)
-import qualified Text.Megaparsec.Char.Lexer        as L
-
--- External imports: auxiliary
-import Data.ByteString.Extra as B (safeReadFile)
+import qualified Copilot.Core as Core
+import           Data.List    (intercalate)
 
 -- Internal imports: auxiliary
 import Copilot.Core.Analysis        (exprIsConstant)
 import Copilot.Language.Reify.Extra (reifySpec)
 import Data.Diagram                 (Diagram (..), diagramNumStates,
                                      diagramStates)
-import Data.ExprPair                (ExprPair (..), ExprPairT (..),
-                                     exprPairShow)
-
--- | Diagram formats supported.
-data DiagramFormat = Mermaid
-                   | Dot
-  deriving (Eq, Show)
-
--- * Diagram parsers
-
--- | Read a diagram from a file.
-readDiagram :: FilePath                    -- ^ File containing diagram
-            -> DiagramFormat               -- ^ Format of the input file
-            -> ExprPair                    -- ^ Subparser for conditions or
-                                           -- edge expressions
-            -> IO (Either String Diagram)
-readDiagram fp format exprP = do
-  contentEither <- B.safeReadFile fp
-  return $ do
-    -- The following functions use Either to return errors. The use of (>>=) to
-    -- chain functions makes the program stop at the earliest error.
-    diagFileContent <- contentEither
-
-    -- Abtract representation of a state machine diagram.
-    parseDiagram format diagFileContent exprP
-
--- | Generic function to parse a diagram.
-parseDiagram :: DiagramFormat          -- ^ Format of the input file
-             -> B.ByteString           -- ^ Contents of the diagram
-             -> ExprPair               -- ^ Subparser for conditions or edge
-                                       -- expressions
-             -> Either String Diagram
-parseDiagram Dot     = parseDiagramDot
-parseDiagram Mermaid = parseDiagramMermaid
-
--- ** Dot parser
-
--- | Parse a DOT / Graphviz diagram.
-parseDiagramDot :: B.ByteString -> ExprPair -> Either String Diagram
-parseDiagramDot contents exprP = do
-    let contentsUTF8 = toUTF8 contents
-    dg <- fst $ G.runParser G.parse $ preProcess contentsUTF8
-    return $ makeDiagram dg
-  where
-    makeDiagram :: Gs.DotGraph LT.Text -> Diagram
-    makeDiagram g = Diagram links
-      where
-        links = map edgeToLink (graphEdges g)
-
-        edgeToLink edge =
-            ( read (LT.unpack o)
-            , exprPairShow exprP (LT.unpack e)
-            , read (LT.unpack d)
-            )
-          where
-            o = G.fromNode edge
-            d = G.toNode edge
-            e = getLabel (G.edgeAttributes edge)
-
-            -- Extract the label from a list of attributes. If no label is
-            -- found, it's assumed that the condition is the literal true.
-            getLabel [] = "true"
-            getLabel ((Attributes.Label (Attributes.StrLabel l)) : _) = l
-            getLabel (_ : as) = getLabel as
-
--- ** Mermaid parser
-
--- | Parse a mermaid diagram.
-parseDiagramMermaid :: B.ByteString -> ExprPair -> Either String Diagram
-parseDiagramMermaid txtDia exprP =
-    case parsingResult of
-      Left e  -> Left (errorBundlePretty e)
-      Right x -> Right x
-  where
-    txt           = T.decodeUtf8 (toStrict txtDia)
-    parsingResult = parse (spaces *> pDiagram exprP) "<input>" txt
-
--- | Type for parser for memaid diagrams.
-type MermaidParser = ParsecT Void Text Identity
-
--- | Parser for mermaid diagrams.
-pDiagram :: ExprPair -> MermaidParser Diagram
-pDiagram  exprP =
-      pGraphDiagram exprP
-  <|> pStateDiagram exprP
-  <|> pSequenceDiagram exprP
-
--- | Parser for a mermaid diagram.
---
--- This parser depends on an auxiliary parser for the expressions associated to
--- the edges or connections between states.
-pGraphDiagram :: ExprPair -> MermaidParser Diagram
-pGraphDiagram exprP = do
-  _ <- string "graph" <* spaces
-  _name <- T.pack <$> manyTill alphaNumChar (char ';')
-  _ <- newline
-
-  transitions <- many (pGraphTransition exprP)
-
-  pure $ Diagram transitions
-
--- | Parser for an edge in a state diagram.
---
--- This parser depends on an auxiliary parser for the expressions associated to
--- the edges or connections between states.
-pGraphTransition :: ExprPair -> MermaidParser (Int, String, Int)
-pGraphTransition ep@(ExprPair (ExprPairT { exprTParse = parseProp })) = do
-  _ <- spaces
-  stateFrom <- many digitChar
-  _ <- string "-->|"
-  edge <- many (noneOf ("|" :: [Char]))
-
-  let x = parseProp edge
-  when (isLeft x) $ fancyFailure $ Set.singleton $
-    ErrorFail $ "Edge property has incorrect format: " ++ show edge
-
-  _ <- char '|'
-  stateTo <- many digitChar
-  _ <- char ';'
-  _ <- newline
-  return (read stateFrom, exprPairShow ep edge, read stateTo)
-
--- | Parser for Mermaid diagrams of type stateDiagram-v2.
-pStateDiagram :: ExprPair -> MermaidParser Diagram
-pStateDiagram exprPair = do
-  _ <- string "stateDiagram-v2" <* spaces
-
-  transitions <- many (pStateTransition exprPair)
-
-  pure $ Diagram transitions
-
--- | Parser for transition label in stateDiagram-v2 mermaid diagram.
-pStateTransition :: ExprPair -> MermaidParser (Int, String, Int)
-pStateTransition ep@(ExprPair (ExprPairT { exprTParse = parseProp })) = do
-  _ <- spaces
-  from <- read <$> many digitChar
-  _ <- spaces
-  string "-->"
-  _ <- spaces
-  to <- read <$> many digitChar
-  _ <- spaces
-  _ <- char ':'
-  _ <- spaces
-  edge <- many (noneOf ("\n" :: [Char]))
-
-  let x = parseProp edge
-  when (isLeft x) $ fancyFailure $ Set.singleton $
-    ErrorFail $ "Edge property has incorrect format: " ++ show edge
-
-  _ <- newline
-
-  pure $ (from, exprPairShow ep edge, to)
-
--- | Parser for Mermaid diagrams of type sequenceDiagram.
-pSequenceDiagram :: ExprPair -> MermaidParser Diagram
-pSequenceDiagram exprPair = do
-  spaces
-  _ <- string "sequenceDiagram"
-  spaces
-
-  conditions <- many (pSequenceTransition exprPair)
-  let transitions = zipWith (\t idx -> (idx, t, idx + 1)) conditions [0..]
-
-  pure $ Diagram transitions
-
--- | Parser for a connection, message or transition in a sequence diagram.
---
--- This parser depends on an auxiliary parser for the expressions associated to
--- the connections or messages between elements.
-pSequenceTransition :: ExprPair -> MermaidParser String
-pSequenceTransition ep@(ExprPair (ExprPairT { exprTParse = parseProp })) = do
-  spaces
-  stateFrom <- many digitChar
-  spaces
-  pSequenceArrow
-  spaces
-  stateTo <- many digitChar
-  spaces
-  _ <- char ':'
-  spaces
-  edge <- many (noneOf ("\n" :: [Char]))
-
-  let x = parseProp edge
-  when (isLeft x) $ fancyFailure $ Set.singleton $
-    ErrorFail $ "Edge property has incorrect format: " ++ show edge
-
-  _ <- newline
-
-  pure (exprPairShow ep edge)
-
--- | Parser for arrow in sequence diagram.
-pSequenceArrow :: MermaidParser ()
-pSequenceArrow = void $ choice
-  [ string "->>"
-  , string "-->>"
-  , string "-)"
-  ]
-
--- | Consume spaces
-spaces :: MermaidParser ()
-spaces = L.space space1 empty empty
 
 -- * Analysis of Specs
